delete Register, fix environment, update VerifiableCredential interface
