--- src/app/pages/register/register.page.html
@@ -1,57 +1,0 @@
-
-<ion-content [fullscreen]="true" class="ion-padding content">
-
-  <br/>
-        <ion-grid class="login-grid">
-          <ion-row>
-            <ion-col class="ion-justify-content-first" style="margin-top:20px">
-              <ion-col class="ion-justify-content-first" style="border-left-width:3px;border-color: #B0FF2B; margin-left: 10px;border-left-style: solid;">
-              <ion-text class="ion-text-login" color="primary"><b>{{"register.title"|translate}}</b></ion-text></ion-col>
-            </ion-col>
-            <ion-col class="ion-text-right header-logo">     
-              <img alt src="assets/icon/IN2-SLOGAN.png" width="auto"  height="60px"/>
-            </ion-col>
-          </ion-row>
-  
-          <ion-row class="ion-justify-content-center">
-            <form [formGroup]="login" (ngSubmit)="onSubmit()">
-              <ion-item class="item-username">
-                <ion-input placeholder="Username" labelPlacement="floating" type="text" required formControlName="username"></ion-input>
-                <ion-icon name="person-outline" slot="start"></ion-icon>
-  
-              </ion-item >
-              <br/>
-              <ion-item class="item-username">
-                <ion-input placeholder="Email" labelPlacement="floating" type="email" required formControlName="email" (focusout)="handleEmailBlur()"></ion-input>
-                <ion-icon name="mail-outline" slot="start"></ion-icon>
-  
-              </ion-item >
-              <ion-item class="email-error" *ngIf=" login.get('email')?.hasError('email') && showEmailError">
-                <ion-label>{{"register.email-error"|translate}}</ion-label>
-              </ion-item>
-              <br/>
-              <ion-item class="item-username">
-                <ion-input placeholder="Password" labelPlacement="floating" type="password" required formControlName="password"></ion-input>
-                <ion-icon name="lock-closed-outline" slot="start"></ion-icon>
-  
-              </ion-item>
-  
-  
-              <ion-row class="login-row">
-                <ion-col>
-                  <ion-button class="ion-button-in2"expand="block"color="primary" type="submit" [disabled]="!login.valid">{{"register.register"|translate}}</ion-button>
-                </ion-col>
-              </ion-row>
-              <ion-row class="ion-justify-content-center">
-                <ion-button class="ion-button-register" fill="clear" color="secondary"[routerLink]="['/login']"><ion-icon name="arrow-back-outline"></ion-icon>{{"register.login"|translate}}</ion-button>
-              </ion-row>
-            </form>
-          </ion-row>
-          <ion-row *ngIf="error" class="ion-justify-content-center">
-            <p class="error-message" translate>{{ error }}</p>
-          </ion-row>
-        </ion-grid>
-  
-    
-  </ion-content>
-
