--- src/app/pages/register/register.page.scss
@@ -1,31 +1,0 @@
-.register{
-    width: -webkit-fill-available;
-}
-.content{
-    --background: url("https://www.in2.es/images/backgrounds/m3@2x.jpg") center/cover fixed no-repeat;
-}
-.item-trns {
-    --ion-item-background: transparent;
-	color: #B0FF2B; 
-}
-ion-item{
-    --border-color:#B0FF2B;
-}
-.custom-button-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-.ion-button-custom{
-    --ion-color-base:#B0FF2B!important;
-    --ion-color-contrast:black!important;
-}
-.ion-button-custom[disabled]{
-    --ion-color-base:grey!important;
-  }
-  .header-logo{
-    height: 9rem ;
-  }
-  .login-row{
-    margin-top: 2rem;
-  }
